feat(info): 将 Release Size 转换成人类可读
