+++ os-checks/shared/columns-select.ts
@@ -73,3 +73,15 @@ export class Cols {
   }
 }
 
+export function formatBytes(bytes: number, decimals = 0) {
+  if (bytes === 0) { return '0 B'; }
+
+  const k = 1024; // 1KB = 1024 bytes
+  const dm = decimals < 0 ? 0 : decimals; // 小数位数
+  const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+
+  const i = Math.floor(Math.log(bytes) / Math.log(k));
+
+  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
+}
+
