+++ os-checks/shared/repos.ts
@@ -71,18 +71,6 @@ export type Output = {
   contributors: Contributor[],
 }
 
-export function formatBytes(bytes: number, decimals = 0) {
-  if (bytes === 0) { return '0 B'; }
-
-  const k = 1024; // 1KB = 1024 bytes
-  const dm = decimals < 0 ? 0 : decimals; // 小数位数
-  const sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-
-  const i = Math.floor(Math.log(bytes) / Math.log(k));
-
-  return parseFloat((bytes / Math.pow(k, i)).toFixed(dm)) + ' ' + sizes[i];
-}
-
 export type Repo = {
   idx: number,
   user: string,
